learningtest.sql updates.

+++ sql/learningtest.sql
@@ -38,6 +38,6 @@ DROP TABLE IF EXISTS cluster_centers;
 CREATE TABLE cluster_centers (
   device_id int,
   cluster_id bigint,
-  longitude float,
-  latitude float
+  longitude double,
+  latitude double
 );
